backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
websockets==12.0

# Image processing
opencv-python==4.9.0.80
numpy==1.26.3
Pillow==10.2.0

# Face recognition
insightface==0.7.3
onnxruntime==1.16.3

# Database
aiosqlite==0.19.0
